Add the first (meta) architectural decision document
Architecture Decision Records (ADR) should be used to document
architectural decisions that we make during development. Since it's a
fact that noone reads long documentation files, having a concise
format is important. Using the source repository has the following
advantages:

* decisions can be reviewed and commented on by the team;
* the decisions are versioned along with the implementation
reflecting them.

Finally, using the repository's web UI, URLs to the master branch can
be used to link to the latest version of any decision.

For more information, please take a look at ADR-0001. As it explains,
the adr CLI is recommended to create and supersede ADRs.