Enable blacklisting of JWTs by their jti claim
In a distributed setting, the node that does the blacklisting spreads
the information via Phoenix' PubSub Server as Phoenix Presence
information. The other nodes react by tracking the same record
themselves, which means that for one record and n nodes there are n
items in the Presence list. The following properties are a result of
this:

- Blacklisting can occur on/by any node.
- The blacklist is eventually consistent over all nodes.
- Any node can go down and come up at any time without affecting the
blacklist, except if all nodes go down at the same time (in that case
there is nothing to synchronize from -- the list is not stored on disk).