Use dedicated Kafka groups to allow multi-purpose Kafka topics
Using the same Kafka group everywhere means that only one of the consumers will
get the Kafka message. However, in RIG each Kafka handler represents a different
feature and they must not affect each other.

For KAFKA_TO_FILTER and KAFKA_TO_HTTP this is actually a bugfix, because both
handlers would typically be subscribed to the same topic, yet in that case only
one of them would receive its messages.

For PROXY_KAFKA_RESPONSE_TOPICS this change allows sharing the Kafka topic with
the other consumers as well. However, since this causes RIG to process each
event at least twice, this is not recommended.